[#1087] Made first HomePage refactor.
